requirements.txt
asgiref==3.4.1
Django==4.0
django-cors-headers==3.10.1
django-filter==21.1
djangorestframework==3.12.4
djangorestframework-camel-case==1.3.0
importlib-metadata==4.8.2
Markdown==3.3.6
Pillow==8.4.0
pytz==2021.3
sqlparse==0.4.2
zipp==3.6.0
